New / Rising / Controversial Sorting
Fun stuff (my first contribution on GitHub)!

In some cases I've wanted to look at the new items in a feed but it
didn't seem like this was available.  It was really easy to add by
adding the values to the sorting select and updating the js to always
apply the sorting to the uri since /hot/ works fine for the standard
feeds.  It seems to work well except sometimes rising glitches out for
some reason - meh.

Also added some minor css tweaks and batman'd the limit to pull from
the Globals in I feel like setting it to 1000 or something crazy the
next time I'm working a marathon :)